BIG COMMIT: 1-ADDITIONAL APPLICATION TO DOWNLOAD MODULES. Added an additional "entrance point" in the launcher (Oppi Store) that will display the contents of an online file (see our wiki for more info)
Now the modules can be donwloaded and installed from that store.

Once the module is installed, it will be displayed in the "main view" for oppi application.

2-DATABASE USAGE TO STORE MODULE POINTERS TO  THE LOCATION OF RESOURCES IN THE STORAGE AREA.
The parsing happens once the user clicks a module in order to play it.
By now, the database has the data requiered to show the icon, name, version, installed path,etc, but not the information to do playback.

3-FULL OVERHAUL OF THE PARSER.
The JSONParser was deleted, now the parsing is done by an static class that uses GSON, which is much cleaner.
The ContentPackage, Frame, ImageFrame, etc, where slightly modified to make it easier to parse directly with GSON.

NOTE: The scene.json file format was modified. See our wiki for information.

Even thow the raw resources with the json files still exist, they are now OBSOLETE.
(To be deleted/replace in a further commit)

5- PLAYBACK IS FAILLING:
There is no sound, and the scene does not change. No bugs files yet, as this still a work in progress.